Moved structs into their own file and tested game fully
