+++ model/game.go
@@ -1,4 +1,4 @@
-package main
+package model
 
 // type Game struct {
 // 	player *Player
