+++ model/item.go
@@ -1,5 +1,12 @@
 package model
 
+type Item struct {
+	Name string
+	Description string
+	Weight int
+	Hidden bool
+}
+
 func (i *Item) SetDescription(description string) {
 	i.Description = description
 }
