+++ model/room.go
@@ -1,5 +1,13 @@
 package model
 
+type Room struct {
+	Name string
+	Description string
+	Exits map[string]*Room
+	Items map[string]*Item
+	Entities map[string]*Entity
+}
+
 func (r *Room) SetDescription(description string) {
 	r.Description = description
 }
